+++ objectified-ui/public/WHATS_NEW.md
@@ -40,6 +40,7 @@ We continue to improve the platform based on your feedback with improvements and
     - Added ability to export the entire viewport, or the current viewport
     - Added ability to export selected groups in the viewport
     - Added ability to include styling elements in the export options
+  - Added ability to delete classes that have been to a group
 - Paths Improvements:
   - Added API Key security schemes: define header, query, or cookie-based API keys in the Security tab and use them on operations (#410)
   - Added inline variable editor that can click parameters in the header of the path studio